requirements.txt
PyQt5>=5.15.0
Pillow>=11.0.0
pywin32>=307
pystray>=0.19.4
keyboard>=0.13.5
